move push_remind method into batch/push_reminds.py
